# cloudsuite-hr-api/src/docs/swagger.yaml
basePath: /
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API para controle de registro de ponto.
  title: Cloudsuite HR API
  version: "1.0"
paths:
  /times:
    get:
      description: Gets a list of all time entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Retrieves all time entries
      tags:
      - Times
    post:
      consumes:
      - application/json
      description: Creates a time entry for an employee
      parameters:
      - description: Time entry details
        in: body
        name: time
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Creates a new time entry
      tags:
      - Times
  /times/date/{date}:
    get:
      description: Gets a list of time entries for a specific date
      parameters:
      - description: Date in the format YYYY-MM-DD
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Retrieves time entries by date
      tags:
      - Times
  /times/day/{day}:
    get:
      description: Gets a list of time entries for a specific day
      parameters:
      - description: Day in the format DD
        in: path
        name: day
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Retrieves time entries by day
      tags:
      - Times
  /times/month/{month}:
    get:
      description: Gets a list of time entries for a specific month
      parameters:
      - description: Month in the format MM
        in: path
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Retrieves time entries by month
      tags:
      - Times
  /times/year/{year}:
    get:
      description: Gets a list of time entries for a specific year
      parameters:
      - description: Year in the format YYYY
        in: path
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Retrieves time entries by year
      tags:
      - Times
swagger: "2.0"